refactor: 접근제어자 변경, 패키지 위치 변경
