--- app/src/main/java/com/techcourse/dao/JdbcTemplate.java
@@ -1,62 +1,0 @@
-package com.techcourse.dao;
-
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-import javax.sql.DataSource;
-
-public class JdbcTemplate {
-
-    private final DataSource dataSource;
-
-    public JdbcTemplate(DataSource dataSource) {
-        this.dataSource = dataSource;
-    }
-
-    public void update(String query, Object... params) {
-        try (Connection conn = dataSource.getConnection();
-                PreparedStatement pstmt = conn.prepareStatement(query)) {
-            setValues(pstmt, params);
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public <T> T queryForObject(String query, RowMapper<T> rowMapper, Object... params) {
-        return query(query, rowMapper, params).get(0);
-    }
-
-    public <T> List<T> query(String query, RowMapper<T> rowMapper, Object... params) {
-        try (Connection conn = dataSource.getConnection();
-                PreparedStatement pstmt = conn.prepareStatement(query)) {
-            setValues(pstmt, params);
-            pstmt.executeQuery();
-            return doMapping(pstmt.executeQuery(), rowMapper);
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public <T> List<T> doMapping(ResultSet resultSet, RowMapper<T> rowMapper) throws SQLException {
-        List<T> data = new ArrayList<>();
-        while (resultSet.next()) {
-            data.add(rowMapper.mapRow(resultSet));
-        }
-        if (data.isEmpty()) {
-            throw new SQLException();
-        }
-        return data;
-    }
-
-    public void setValues(PreparedStatement pstmt, Object... params) throws SQLException {
-        int parameterNumber = 1;
-        for (Object param : params) {
-            pstmt.setObject(parameterNumber++, param);
-        }
-    }
-}
