+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -3,6 +3,8 @@ package com.techcourse.dao;
 import com.techcourse.domain.User;
 import java.util.List;
 import javax.sql.DataSource;
+import nextstep.jdbc.JdbcTemplate;
+import nextstep.jdbc.RowMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
